# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sondhan AI - Lost & Found Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /lost-items:
    get:
      operationId: listLostItems
      tags: [lost-items]
      summary: List lost items
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of lost items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LostItem"
    post:
      operationId: createLostItem
      tags: [lost-items]
      summary: Report a lost item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LostItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
  /lost-items/{id}:
    get:
      operationId: getLostItem
      tags: [lost-items]
      summary: Get lost item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lost item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLostItem
      tags: [lost-items]
      summary: Update lost item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LostItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
    delete:
      operationId: deleteLostItem
      tags: [lost-items]
      summary: Delete lost item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /found-items:
    get:
      operationId: listFoundItems
      tags: [found-items]
      summary: List found items
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of found items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoundItem"
    post:
      operationId: createFoundItem
      tags: [found-items]
      summary: Report a found item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoundItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoundItem"
  /found-items/{id}:
    get:
      operationId: getFoundItem
      tags: [found-items]
      summary: Get found item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoundItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateFoundItem
      tags: [found-items]
      summary: Update found item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoundItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoundItem"
    delete:
      operationId: deleteFoundItem
      tags: [found-items]
      summary: Delete found item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        universityId:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        batch:
          type: ["string", "null"]
        role:
          type: string
        phone:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isVerified
        - createdAt
...
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        department:
          type: string
        batch:
          type: string
        profileImage:
          type: string
    LostItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: string
        image:
          type: ["string", "null"]
        dateLost:
          type: ["string", "null"]
        lastSeenLocation:
          type: ["string", "null"]
        status:
          type: string
        reportedBy:
          type: integer
        reporterName:
          type: ["string", "null"]
        aiMatches:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - description
        - status
        - reportedBy
        - createdAt
    LostItemInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        image:
          type: string
        dateLost:
          type: string
        lastSeenLocation:
          type: string
      required:
        - title
        - category
        - description
    LostItemUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        image:
          type: string
        dateLost:
          type: string
        lastSeenLocation:
          type: string
        status:
          type: string
    FoundItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: string
        image:
          type: ["string", "null"]
        dateFound:
          type: ["string", "null"]
        foundLocation:
          type: ["string", "null"]
        status:
          type: string
        foundBy:
          type: integer
        founderName:
          type: ["string", "null"]
        finderContact:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - description
        - status
        - foundBy
        - createdAt
    FoundItemInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        image:
          type: string
        dateFound:
          type: string
        foundLocation:
          type: string
        finderContact:
          type: string
      required:
        - title
        - category
        - description
    FoundItemUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        image:
          type: string
        dateFound:
          type: string
        foundLocation:
          type: string
        status:
          type: string
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
...