# lib/api-zod/src/generated/types/lostItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sondhan AI - Lost & Found Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LostItemInput {
  title: string;
  category: string;
  description: string;
  image?: string;
  dateLost?: string;
  lastSeenLocation?: string;
}